feat: Add memory circuit breaker and kill slow query (#699)

+++ internal/engine/search/engine.h
@@ -135,7 +135,7 @@ class Engine {
 
   int AddNumIndexFields();
 
-  int MultiRangeQuery(Request &request, SearchCondition *condition,
+  int64_t MultiRangeQuery(Request &request, SearchCondition *condition,
                       Response &response_results,
                       MultiRangeQueryResults *range_query_result);
